[evcc] Address comments
* Remove catching NullPointerException
* Remove catching of Exceptions in refresh methods as there normally should never be no exceptions.

+++ bundles/org.openhab.binding.evcc/src/main/java/org/openhab/binding/evcc/internal/dto/Loadpoint.java
@@ -40,7 +40,7 @@ public class Loadpoint {
     private long chargeRemainingDuration;
 
     @SerializedName("chargeRemainingEnergy")
-    private long chargeRemainingEnergy;
+    private double chargeRemainingEnergy;
 
     @SerializedName("chargedEnergy")
     private double chargedEnergy;
@@ -149,7 +149,7 @@ public class Loadpoint {
     /**
      * @return the chargeRemainingEnergy
      */
-    public long getChargeRemainingEnergy() {
+    public double getChargeRemainingEnergy() {
         return chargeRemainingEnergy;
     }
